[MIG] stock_report_quantity_by_location: Migration to 16.0
