+++ stock_report_quantity_by_location/__manifest__.py
@@ -3,7 +3,7 @@
 {
     "name": "Stock Report Quantity By Location",
     "summary": "Stock Report Quantity By Location",
-    "version": "15.0.1.0.0",
+    "version": "16.0.1.0.0",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/stock-logistics-reporting",
     "category": "Warehouse Management",
