+++ stock_report_quantity_by_location/wizards/__init__.py
@@ -1,1 +1,2 @@
+from . import stock_report_quantity_by_location_prepare
 from . import stock_report_quantity_by_location
